requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.6
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-import-export==4.3.7
psycopg2-binary==2.9.10
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
